requirements.txt
# 核心檢測與OCR
manga-ocr>=0.1.11
torch>=2.0.0
torchvision>=0.15.0

# 圖像處理
opencv-python>=4.8.0
Pillow>=9.5.0

# Google Gemini API
google-generativeai>=0.3.0

# 基礎依賴
numpy>=1.24.0
pathlib
argparse

# 其他依賴
tesserocr>=2.6.2
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0.0
scikit-image>=0.21.0 